requirements.txt
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
backoff==2.2.1
beautifulsoup4==4.13.3
browser-use==0.1.40
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
defusedxml==0.7.1
distro==1.9.0
fastapi==0.115.12
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-anthropic==0.3.3
langchain-core==0.3.50
langchain-ollama==0.2.2
langchain-openai==0.3.1
langsmith==0.3.24
markdownify==0.14.1
monotonic==1.6
ollama==0.4.7
openai==1.70.0
orjson==3.10.16
packaging==24.2
playwright==1.51.0
posthog==3.23.0
pydantic==2.11.2
pydantic_core==2.33.1
pyee==12.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.1
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
uvicorn==0.34.0
zstandard==0.23.0
